Mark converter trait as unsafe
Clippy correctly complains that the try_from_mrb function should be marked as
unsafe because we are potentially dereferencing the mrb pointer like we do in
the Array macros. This is inherently unsafe because nothing stops mrb from
being a null pointer.

This commit converts all tests to be wrapped in unsafe blocks instead of down
scoping the unsafe.