Fix a massive memory leak in Mrb::eval
mruby has a built in garbage collector. All objects are stored in an
"arena" as they are created. The GC takes the conservative approach
of marking all objects created via C APIs with `mrb_gc_protect` which
tells the GC they are permanently alive.

Mrb::eval uses `mrb_load_nstring_cxt` and gets back an `mrb_value`.
This value is protected so mruby will never garbage collect it. This
means that each call to eval leaks memory because the `mrb_value` is
never freed from the mrb heap by the GC.

This problem is exacerbated by the old `mrb_sys_get_current_exception`
C extension function implemented by `mruby-sys` which calls several
methods on an exception `mrb_value` via `mrb_funcall`.

This commit introduces two tests to ensure that calls to eval are not
leaking memory in the happy path (no raised exception) and sad path
(eval ended with an exception raised on the interpreter) cases.

To fix the leak, we track the current position in the arena before
calling any mrb C API functions, restore the arena to this position
after we have extracted the `mrb_value`, and then call a full GC.

This commit moves the implementation of `mrb_sys_get_current_exception`
to Rust to allow collecting of the gc protected value after converting
it to a native Rust String.