Set filename in mrb context
This commit introduces an EvalContext which can modify how code is loaded into
the mruby interpreter with eval. The EvalContext currently exposes a filename
property which sets the __FILE__ magic constant when running code via
mrb_load_nstring_cxt.

The State tracks a stack of EvalContexts which allows the filename to be persistent
across Rust and C execution. For example, when requiring a Rust type that implements
MrbFile into the mruby interpreter, a context with the filename is pushed onto the
State. If the Rust code evals code on require, the __FILE__ magic constant will be
set to the source location.

__FILE__ is set to the absolute path of the source file in the VFS attached to the
State.

This commit moves the eval method to its own trait like I have been doing with
MrbLoadSources and GarbageCollection. This commit introduces a new eval function
that will simultaneously push an EvalContext onto the stack and execute an eval.

If no EvalContexts are present on the stack, eval will fall back to EvalContext::root
which sets the filename to '(eval)'.

Now that mruby crate sets filenames on the context, MrbApi::current_exception includes
backtraces in the error it returns.

This patch does not set __FILE__ for code evaled by Rust functions that define Ruby
methods.