Allow mutability of class and module specs
Previously, class::Spec and module::Spec objects were stored in the
State in a Map wrapped in Rc smart pointers. The Rc pointers allowed
Specs to have multiple owners, which allowed them to be used for
parent class/module and super class relationships. Rc has enforces that
there is EITHER mutliple RO copies OR one mutable copy of the Rc. This
meant that once a spec was used as a parent or super class, it was
frozen. Attempting to mutate the spec would cause a panic.

The dance to get a mutable reference out of the Rc also made the API
awkward: first call def_class and then call class_spec_mut to get a
mutable reference. THEN call class_spec to get a normal reference to
allow cloning. All of this required carefully managing scopes to prevent
panics from the Rc borrow mechanics.

This commit modifies the spec storage in State to store specs in an
Rc<RefCell<_>>. This provides interior mutability which simplifies use
of the smart pointer returned by class_spec at the cost of callers having
to borrow/borrow_mut. In practice this makes code that manipulates specs
much simpler, as shown in manual.rs.

To make this possible, I had to implement Hash manually for Parent
instead of deriving it by borrowing and delegating to the inner spec.

TODO: Need a better story for generating "fake" specs for builtin Ruby
types like StandardError.