Return Err from eval if code does not parse
There are some cases, like when code does not parse, where mrb_load_nstring_cxt
returns mrb_undef_value. undef is meant to never be exposed to the Ruby VM and
doing so is undefined behavior that may result in a segfault.

mruby crate already has a concept of an unreachable mrb_value. This commit
detects if the value we extract from the the interpreter with eval is unreachable
and returns an error true.

See mruby/mruby#4460.