Reimplement array converters with infallible FromMrb trait
This commit implements Vec and slice converters for mruby primitive
types of boolean, bytes, fixnum, float, and string. Collection types
array and hash are excluded.

This commit implements each type in code instead of using a macro. This
improves readability and is easier to maintain and reason about.

It also removes the ridiculous Vec of Vec implementations. These impls
are unlikely to be useful and removing them speeds up test execution
significantly.