Fix leak when creating Rack app
adapter::from_rackup creates a Value that is a Rack app with the C API
which requires restoring the arena to its previous position. We were
already managing the arena in the prefork route, but were only saving
the position after creating the Rack app.

This commit also restores the arena position in handler::run after
creating a Rocket response.