Add support for a thread-local app specific interpreter
This reimplements support for the prefork exec mode previously in foolsgold
crate. Shared interpreters can be recycled with the max_requests parameter
which has the same effect as the max_requests parameter in gunicorn: for each
max_requests, tear down the interpreter and lazily initialize a new one on the
next request the worker receives. This parameter exists to mitigate memory
leaks.