Protect against stack splits when raising from Rust code
sys::mrb_raise is implemented with longjmp, which modifies the stack.
Rust callers do not expect this and it was causing mutable borrow
errors on Mrb. To fix this, wrap calls to raising C APIs (eval, raise)
with sys::mrb_protect and sys::mrb_ensure.

With this fix, nested calls to eval and raise, such as 'eval raise',
do not cause a panic.

This commit also re-syncs the implementation of mruby VM interactions
across funcall, eval, and raise.