Add a benchmark for String#scan
Based on the benchmark found here:

    https://github.com/mariomka/regex-benchmark/tree/master/ruby

The benchmark looks for emails, URIs, IPs, in a corpus derived from
"Learn X in Y Minutes", which is about 6MB and contains a range of
UTF-8 characters.

The benchmark is runnable with the mruby crate using a custom runner
found in mruby-bin. To invoke the benchmark, run:

    cargo run --release --bin string_scan_bench -- mruby-bin/ruby/benches/string_scan.rb

The benchmark is runnable with CRuby. The .ruby-version file in
ferrocarril specifies Ruby 2.6.3. To invoke the benchmark, run:

    ruby mruby-bin/ruby/benches/string_scan.rb