Remove MrbApi trait
These factory methods for Values dangling off the Mrb interpreter are
an artifact of a more immature codebase. Back then, the only converters
were TryFromMrb which were both falible and unsafe. These factories
existed to unwrap and hide the unsafety.

FromMrb is a safe, infalible converter which has existed for a while.
Use of FromMrb and MrbApi factories was split about 50-50, so this
commit nukes the more specialized interface.

This removes the inconsistency where only primitive factories were
exposed, i.e. could not create an Array or Hash.