Add unchecked eval methods that allow the stack to unwind
When Kernel#require calls eval, it is not at the top-level so it does not want
to catch exceptions, which is what mrb_protect in eval does. Implement
unchecked_eval which calls mrb_load_nstring_cxt directly. This allows longjmp
to unwind past the eval call in require to the top-level protect.

Fixes GH-154.