Fix memory leak in RubyException::raise
mrb_raise is implemented with setjmp/longjmp. Calling raise from Rust is
super unsafe because unwinding the stack does not give Rust the the
chance to clean up stack frames and ensure that heap memory is
reclaimed.

The fix is to have raise consume its arguments and drop everything
before calling mrb_raise. It is only safe to call RubyException::raise
from Rust code that is one frame below the mruby VM. Unwinding past a
Rust stack frame will result in a leak. This means that unchecked_eval
must be used with extreme caution! A future patch will mark this
function as unsafe.

As a follow up, I plan on removing all of the macros in macros.rs as it
seems important to retain explicit control over when raise is called.

The following code snippets demonstrate the leak:

    # exception (TypeError) during Regexp.new
    pat = ['a'] * 1024 * 1024; (4096 * 512).times { Regexp.new(pat) rescue nil }
    # exception (IndexError) during MatchData#[]
    group = 'a' * 1024 * 1024; (4096 * 512).times { /./.match('a')[group] rescue nil }

The MatchData leak is particularly bad, ballooning the REPL to over 25GB
of RSS very quickly.

TODO: add a regression test.

Fixes GH-165.