Ensure code is deallocated at the end of MrbEval::eval
Calls to mrb_load_nstring_cxt potentially unwind with longjmp. If the only
reference to the code is in an owned Rc in run_protected, the unwinding can
leak the code Vec.

Clone the Rc before passing it to protect and deallocate the args Rc including
code buffer adter the call to mrb_protect.

This commit refactors run_protected out of the eval definition in preparation for
making the same fix to unchecked_eval.