Prevent String leaks in RubyException::raise
I discovered in mruby/mruby#4558 that mrb_raise is only meant to be
called with static strings. This is because mrb_raise unwinds the stack
with longjmp and there is no opportunity to deallocate a heap-allocated
string.

Change the type bound on the message parameter to RubyException::raise
to be &'static str.

Introduce RubyException::raisef which takes a vec of things that can be
converted to Value. mrb_raisef takes a va_args of sys::mrb_value which
works around the heap allocation issue by utilizing the mrb heap which
will eventually get garbage collected.

The only users of RubyException::raisef right now are LoadError and
IndexError, each of which only require one format arg. For now, raisef
panics if given more than one format arg.