Use mem::MaybeUninit::uninit to fix deprecation warnings
Replace most uses of mem::uninitialized with mem::MaybeUninit::uninit().
This compiler intrinsic was introduced in Rust 1.36.0.
mem::uninitialized was deprecated in Rust 1.38.0 due to UB.

Most uses of mem::uninitialized were straightforward ports to the new
API. There was one use of mem::uninitialized to create a Default impl
for Regexp that was super lurky.

Convert the Rc<Regex> field to Rc<Option<Regex>>. This was relatively
straightforward to change. If the Option is None, the Regexp is
uninitialized and it makes sense to return Error::Fatal from the method
impl.