Fix bad access when writing *const T into RcBox in Rc::from_raw
Back when I was having trouble getting the unsized bits to compile, I made
the value field on RcBox a Box<T>, which is a pointer to a T. Rc::into_raw
returns a pointer to T that is not the same as the Box. Attempting to write
the *const T into the Box causes a segfault.

Now that the unsized bits are figured out, remove the Box and everything works,
as verified by the doc tests.