Fix segfault in Rc::allocate_for_ptr
Rc::allocate_for_ptr allocates a raw region of memory with the layout of
an RcBox and initializes it by hand. Compared to std::rc, the RcBox in
cactusref has an extra field: a RefCell<Links<T>>. the memory pointed to
by this field was uninitialized which lead to undefined behavior during
Drop when Rc checks if it has any links.