Free Strings from Rust heap after evaling Ruby
All Strings consumed by the interpreter need to be owned by Rust. The Wasm
frontend returns 'pointers' to Strings in a 'heap' (HashMap) and exposes
an API for finding the length of the String behind a pointer and reading
and writing characters from and into the String.

JS copies code into the heap and copies results out of the heap. Once this
eval and result extract is done, the Rust copies of the Strings are no longer
required. Currently, these Strings leak. This PR explicitly frees the Strings
from the heap which will let them be reclaimed by Rust.