Remove dependency on path_abs crate by forking and vendoring
path_abs does not compile under wasm32-unknown-unknown because it
includes conditional compilation on OS vendor targeting unix and
windows. Artichoke does not make use of this functionality.

path_abs canonicalizes relative paths using env::current_dir. Artichoke
uses a virtual filesystem so this behavior is never what we want.

This commit removes the dependency on path_abs and replaces it with a
vendored copy of the PathAbs implementation. This fork modifies the
public API of PathAbs::new to take the "canonicalized" current working
directory to use for absolutizing relative paths.

Tests from upstream are included.

This commit forks path_abs from commit 8370838.