Extract core interpreter traits out of artichoke-backend
This commit extracts all of the behavior-defining traits and converter
traits from artichoke-backend (currently implemented with mruby) to a
base crate.

The intent is to move `extn` into artichoke-core such that core and
stdlib are generically implemented for any Artichoke interpreter that
implements the core traits.

This commit is a start, but there are still some outstanding items:

- RustBackedValue is not yet pulled into core.
- Args extraction and argspecs are not pulled into core.
- Need to figure out how to handle arg parsing for native Rust
  extensions
- Need to figure out how to handle gluing native Rust extensions to
  mruby interpreter with C function pointers.