Update to onig 5.0
onig 5.0 uses bindgen to generate bindings for oniguruma and is easier
to hack on. To upstream the changes in GH-172 that add
wasm32-unknown-unknown support to Artichoke, we need to build with onig
5.0.

In addition to updating the dependency, this commit updates the `Regexp`
and `MatchData` implementations to work with the new APIs exposed by
onig.

The only change is the removal of `Regex::capture_names` and the
addition of the `Regex::foreach_name` function that replaces it. This
changes the API from an iterator to a callback with a break parameter.
This change appears to be a result of upstream changes in oniguruma.

This commit does not include any of the wasm support present in our
fork: https://github.com/artichoke/rust-onig/tree/wasm3. This means that
once this commit is merged, artichoke will no longer build under
wasm32-unknown-emscripten. This is safe because the playground is pinned
to a specific version of artichoke-backend.