Implement Integer#chr
Integer#chr takes an optional Encoding as an argument. Artichoke does
not yet implement Encoding, so this case is unimplemented.

When no Encoding is given, the encoding of the resulting String is
either ASCII or ASCII-8BIT, depending on the value of self. These
constraints on the encoding mean that self must be convertable to
u8. This byte is turned into a single byte Vec<u8>.