Do not use associated types for converter traits in artichoke-core
Associated types are not generics and do not enable multiple
implementations of the trait.

Because we want a separate impl of the converter for every input/output
pair, we need to use real generics.

Convert and TryConvert are now Function1-type traits that are expected
to be implemented on an interpreter type. This avoids the problems we
were having with orphan rules when implementing the traits on types in
std.