Implement artichoke-backend converters with artichoke-core
Reimplement the converters in artichoke-backend to use the traits
defined in artichoke-core. This is a necessary refactor as we pull apart
the Artichoke core from the mruby backend implementation.

The converters are the glue between Ruby and Rust. This PR is messy. It
touches almost every file in artichoke-backend.

This commit replaces `artichoke_backend::ArtichokeError` with
`artichoke_core::ArtichokeError`. `artichoke_backend` reexports
`ArtichokeError` for compatibility and to ease the code migration.

This commit replaces `artichoke_backend::types::{Ruby, Rust}` with the
equivalents in `artichoke_core` and reexports them for compatibility and
to ease the code migration.

This commit reexports `artichoke_core::convert::{Convert, TryConvert}` at
their old locations in `artichoke_backend::convert` for compatibility
and to ease the code migration.

This commit uses macros to implement the `Option`, `Vec`, `Vec<(_, _)>`
and `HashMap<_, _>` container types. Using a macro makes it easier to
get coverage for collection types. As of this commit, artichoke-backend
generates 988 `TryConvert` impls.

Reimplementing the converters with those in artichoke-core made it easier
to see that the converters can be implemented for slices in both directions.
This makes the code simpler and eliminates some allocations.

TryConvert::convert is no longer an unsafe function.

This commit moves callsites to use `value.try_into<_>()` wherever
possible.

To accommodate `&str` conversion, converting a Symbol no longer calls
`Value::to_s` and instead calls the mruby C API directly.

Fixes GH-236.