Properly extract symbol name as raw bytes
mruby does not have an API for getting the raw byte contents of a Symbol name.
For ASCII symbols, this works fine. For non-ASCII names or names that are not
a valid mruby identifier, mrb_sym2name returns a char * with non-ASCII
characters replaced by escape sequences and surrounded by quotes.

Work around this by converting the Symbol to a String and extracting the
byte content from the String. This involves an extra allocation and a memcpy
but it is the best we can do.