Unify ValueLike::funcall and ValueLike::funcall_with_block
Unify funcall and funcall_with_block into ValueLike::funcall that takes
an optional block.

This commit is made possible by and is a follow up to GH-248. This
commit is a prerequisite to implementing Value in artichoke-backend with
the Value trait in artichoke-core.

Simplify the signature of funcall to remove AsRef generics. The method is
now always an &str and the args always a &[Value].

This commit updates every funcall callsite in Artichoke. Update converted
return type to use &str converters to avoid allocations where possible.