Add a macro for extracting arguments from an mruby interpreter
mruby calls `mrb_raise` which unwinds via longjmp if the args passed to
a function do not match the spec passed to mrb_get_args instead of
validating the argspecs declared on function register before delegating
to our Rust implementation. See mruby/mruby#4688.

This is not ideal because unwinds via longjmp do not let Rust clean up
resources (like `RefCell` borrows). This causes bugs like GH-243.

This commit introduces a macro that extracts arguments from the
mrb_state for a predefined set of argspecs. This code uses the argc
returned from `mrb_get_args` to determine which `MaybeUninit`s are
initialized.

This macro is intended to be called before unwrapping the interpreter
from the mrb_state. At this point all stack variables are `Copy` so
unwinding is safe.

Fixes GH-200.