Fix segfault caused by improper `Clone` impl for Array
`mrb_value` is `Copy`. `mrb_value` contains a `*const RBasic`. The
pointer is `Copy`, but the `*const RBasic` is a shared pointer similar
to an `Rc`.

The existing `Clone` implementation calls
`RustBackedValue::try_into_ruby` with a `None` argument for the existing
`mrb_value`. This means that calling `Value::clone` returns a newly
allocated `mrb_value` with a different `RBasic` pointer that happens to
refer to the same `Rc` pointer wrapping the `Array`.

`artichoke_ary_ref` is called during the mruby GC mark phase to color
`mrb_value`s gray which means they are reachable by a GC root. Because
of Rust borrowing rules, indexing into the `Array` beghind the
`mrb_value` to return an `mrb_value` requires a call to `Clone::clone`.

The bug in the clone implementation caused `artichoke_ary_ref` to return
a new `mrb_value` as opposed to a copy of the existing `mrb_value` with
an incremented `Rc`. So when `ARY_REF` was called in the mruby GC to
mark the items in the `Array`, it was actually marking new `mrb_values`
instead of the ones contained in the `Array`. The values in the `Array`
remained white, were freed, and caused segfaults stemming from a
use-after-free error.

The fix is to pass the existing `mrb_value` into
`RustBackedValue::try_into_ruby` to reuse the `RBasic` pointer to ensure
the `RBasic` is marked gray during the GC mark phase.