Fix use-after-free
artichoke_backend::interpreter sets the user data of the mrb_state to an Rc pointer
and returns the Rc without increasing the ownership count. This means that when we
close the interpreter, we should not turn the raw pointer back into an Rc to be
dropped. Dropping the Rc is handled when dropping the external binding to the
interpreter.

It is sufficient to call mrb_close to free all mrb_values on the heap.

ASAN was super useful in tracking this down:

    RUSTFLAGS="-Z sanitizer=address" cargo +nightly test --target x86_64-apple-darwin -p artichoke-backend rust_backed_mrb_value_smart_pointer_leak -- --nocapture