Replace Array implementation with a class cluster
This commit extends and replaces the existing `VecDeque` backed `Array`
implementation with a cluster of 7 concrete `ArrayType`s.

The basic type is `Buffer`, which uses a `Vec` for storage.

There are several specialized types for `Array`s with fixed contents:

- Empty
- One element
- Two element
- Hole, a run of `nil`s
- Repeated runs of a single element

The `Array` struct that gets serialized into an `mrb_value` is a new
type wrapper struct around a `Box<dyn ArrayType>`. `Array` implements a
class cluster similar to `NSArray` in Foundation SDK via dynamic
dispatch on trait objects.

Not all `ArrayType`s are growable, so mutating methods accept a
reallocation out parameter. The public facade handles reallocations by
delegating to an `Aggregate` struct that is a `Vec` of `ArrayType`s.

Some properties that the Artichoke `Array` has that MRI does not:

- Constant time concat
- Partial reallocation of aggregates
- Constant time reverse, and insertion in some circumstances
- Allocation-free mutations in some circumstances
- Sparse arrays
- Arrays larger than `usize::MAX` elements

    artichoke 0.1.0 (2019-10-29 revision 1895) [x86_64-darwin]
    [Rust 1.38.0 (rev 625451e) on x86_64-apple-darwin]
    >>> a = []
    => []
    >>> a[100_000_000_000] = 'end'
    => "end"
    >>> a[0, 0] = Array.new(200_000_000_000, 'repeated'); nil
    => nil
    >>> a.length
    => 300000000001
    >>> b = a.reverse; nil
    => nil
    >>> b.first(10)
    => ["end", nil, nil, nil, nil, nil, nil, nil, nil, nil]