Import prototype of Regexp implementation using regex crate
I wrote this code back in the beginning of the summer as a proof of concept
and never got it merged. This is the code as it was sitting in my Desktop.

This PR exists to get the code off of my machine. The code is not wired
up and is not compiled as part of the build. The patterns in the codebase
have changed considerably since this code was written.

The way forward is to turn `Regexp` into a newtype wrapper around a
`RegexpType` trait object.