Protect temporaries from GC in mruby VM
There are portions of the mruby VM where `Array`s are manipulated, for
example to set up arguments when calling a function. To facilitate the
rollout of the Artichoke Array within mruby, I introduced macros that
allow indirection between the mruby Array and Artichoke Array
implementations.

The mruby implementations mostly operate on a raw buffer via the
`RARRAY_PTR` macro. Because accesses are pointer arithmetic and stack
copies, there is no allocation in the VM.

The Artichoke implementations reuse the same functions for implementing
the Ruby `Array` API. These functions _can_ allocate Ruby objects.

If a full GC takes place while one of these VM temporaries is alive, the
VM state is invalid and we enter UB territory.

Additionally, there are parts of the mruby VM that natively operate on
raw buffers that were rewritten to generate a temporary buffer from an
Artichoke Array.

In both cases, all of the temporaries need to be protected from the GC
by calling `mrb_gc_protect` _before_ calling any functions that can
allocate.

This bug was nasty and only showed up when the allocation order was
tickled just right for extended program executions.

Thanks to @tummychow for help debugging this one.