Implement Regexp with multiple backends and a boxed trait object
This patch changes the implementation of `Regexp` to have multiple
backends that are delegated to with a boxed trait object. This is the
same pattern we are using elsewhere in `Array` and `ENV`.

There are two backends:

- `Onig` is a port of the existing implementation.
- `Lazy` allows constructing a `Regexp` from an already parsed pattern
  without compiling a new `onig::Regex`, This is useful for cases where
  a `MatchData` must be constructed before entering a block but may not
  be accessed.

`MatchData` is modified to accept a `Box<dyn RegexpType>`. All
manipulations on a `Regexp` are pulled out of `MatchData` into the
`RegexpType` interface.

`RegexpType` has an idealized interface where all of the patterns and
haystacks are byte slices instead of `&str`. Concrete implementations
raise `ArgumentError` if they can not handle a pattern or haystack.

This patch passes 12 more specs than the old implementation. While
working on this patch, I added a missing spec to ruby/spec, see
ruby/spec#727. The implementation does not currently pass this new spec.
I will address it when we revendor ruby/spec.

This commit paves the way for implementing the regex crate backend, see
GH-325.