Null out the mruby host build
mruby requires a host build that is not configurable enough to build
Artichoke. This has meant that Artichoke builds the host build to
bootstrap a 'sys' crossbuild that is used to produce the static library.

In the beginning, mruby's build process was used directly to produce
the static library that build.rs would link into Artichoke. More
recently, the sys crossbuild would only do C source generation and
build.rs would discover and compile these sources with the cc
crate.

Now that Artichoke no longer depends on mrblib or Ruby sources in
mrbgems, mrbc is no longer required and the host build is no longer
needed.

This PR replaces the host build with a noop. This changes the code
emitted by the mruby build process to eliminate gem irep embedding.
The bootstrap gembox no longer includes or builds any mrbgems.

Fixes GH-303.