Reduce allocations of Symbols backed by &'static byte strings
The `Symbol` cache in `State` held owned `String`s as the keys.
This was suboptimal because:

1. Symbols can contain arbitrary byte content.
2. Most symbols are known at compile time and have &'static lifetime.

This PR changes the type of the keys in the `Symbol` cache to
`Cow<'static, [u8]>`. This allows the state to avoid allocating a byte
vector unless it needs to to satisy ownership constraints.

Fixes GH-317.