Remove artichoke-core 'exception'
artichoke-backend has developed a nice notion of RubyExceptions with
boxed trait objects and raising with impl trait. The exceptions in
artichoke-backend are coupled to the interpreter for initialization.
They _could_ be split out into artichoke-core but I'm not sure what
the benefit of doing so would be.

The current split is:

- Traits in artichoke-core.
- Data (structs) in artichoke-backend.