Remove `Rc` from class and module spec API
The existing API around class and module specs stores specs in a
type-keyed `HashMap` in the state with specs wrapped in `Rc` pointers.

`Rc` meant that specs required `RefCell` to be mutable. Additionally, to
ensure the `State` had ownership of the spec, there was an awkward dance
in the API to first register the spec with the `State` and then retrieve
it to manipulate it.

This commit changes the API to store raw specs in the `HashMap`, return
regular references in the retrieval APIs, and add `Clone` to the specs
to enable pieces like the `EnclosingRubyScope` to function properly.

Tests do not pass yet.

This commit removes vendored stdlib packages in the hopes of unifying
them to use the auto importer. This change is undone in a later commit
in this PR.