Wrap specs in Box when storing in State
Because references to specs are passed into the mruby VM, the specs must
be wrapped in a Box to ensure their addresses are stable.

This is a regression from the start of this branch where we moved away
from storing Rc in the State, which gave us this property as well.