Handle uninitialized objects in RustBackedValue::try_from_ruby
`Object#allocate` is a Ruby API that creates uninitialized objects.
These objects have the correct class, but have not had the class's
`#initialize` method called on the allocated instance.

In the case of `MRB_TT_DATA`, this means that objects have the correct
class, but they do not have a valid `Rc` stored in the `DATA_PTR` and
they do not have a valid `mrb_data_type`.

Since GH-398 was merged, calling `RustBackedValue::try_from_ruby` with
an object created via `#allocate` would cause a borrow on the
interpreter to not get cleaned up because mruby would unwind past it.
This bad state was surfaced in the spec suite as a `AlreadyBorrowed`
panic when attempting a mutable borrow on the `Artichoke` interpreter.

This commit changes which APIs `RustBackedValue` calls when extracting
pointers from `mrb_value`s with type `MRB_TT_DATA`.

The existing code called the `mrb_data_get_ptr` API, which validates
that the `mrb_data_type` of the target matches the supplied data type.
If there is a mismatch, the function calls `mrb_raise` with a
`TypeError`.

Prior to GH-398, getting a reference to the data type for calling this
API called `interp.0.borrow().class_spec::<Self>()`, which returned a
clone of the `Rc`-wrapped `class::Spec`. Because the returned spec was
an owned `Rc`, the borrow on the interp was implicitly dropped after its
last use. Because the borrow was dropped, the hidden unwind from the C
API did not leave any dangling borrows on the interpreter, but it likely
caused some memory to leak.

The new code introduced in this commit calls `mrb_data_check_get_ptr`
which returns `NULL` instead of raising if the data type is not
validated. Since a `NULL` is now a valid possible return type,
`RustBackedValue::try_from_ruby` returns a new
`ArtichokeError::UninitializedValue` variant so calling trampolines can
return the appropriate `TypeError` to the unsafe C trampoline.

`RustBackedValue::try_from_ruby` is not normally called in an
`mrb_protect` context and the `Artichoke` interpreter is already
extracted. It is not safe to unwind from `RustBackedValue`.

`RustBackedValue` should never should have worked with partially
initialized Ruby values, but did so on accident. This fragile state was
made an error with chnges to borrowing semantics on the interpreter with
the API changes to `class::Spec` in GH-398.