Simplify InlineBuffer implementation
This change was motivated by simplifying the memory layout of
`InlineBuffer` but ended up also reworking some of the core buffer
manipulation algorithms.

Turning `InlineBuffer` into an enum simplified handling the different
states which made it more clear which specialized APIs on `Vec` and
`ArrayVec` to make use of.

The switch to an enum and structuring code with match expressions
revealed some missing branches.

The resulting code makes heavy use of the `extend` family of APIs. There
is no more unsafe code for manipulating the `ArrayVec` outside of the
already unsafe `InlineBuffer::set_len`.

This commit fixes an off by one bug when detecting whether a slice can
be stored in the inline variant.

The new code makes more consistent use of `copied` iterators for
operating on slices.

Initial testing shows no significant performance changes when running
the spec suite, but the suite is dominated by `Regexp` performance. The
gains in correctness and simplicity afforded by this change make it the
right path forward.

Closes GH-414.