Eagerly parse a Context filename into a CString
The API for constructing a `Context` allowed creating Contexts that
could never be used as Contexts. `Context::new` takes effectively a byte
slice and is infallible. `Context` is used as a `CString` by
`Artichoke`'s `Eval` implementation to set the filename on the mruby
parser.

This restriction means that the byte slice used to make the Context
cannot contain any NUL bytes. This restriction also means that getting a
`CString` from the `Context` now must be fallible. Falliblilty when
getting the `Context`'s `CString` is inconvenient in APIs like
`unchecked_eval` which themselves are infallible.

This PR alters the constructor `Context::new` to return
`Option<Context>`. This changes `Context` so that it parses the
`CString` on create instead of validating at use whether the object is
valid.

This change produces better code because callers of `Context::new` can
handle errors locally at the time they are encountered instead of silent
corruption later on. A good example of this is the require code, which
it turns out already has to eforce variants around NUL bytes to be
ruby/spec compliant.

For cases where filenames are static and known to the interpreter at
compile time (like the `-e` or `(eval)` names), an unsafe function
called `Context::new_unchecked` can be used.