Propagate errors from parsing Ruby method names into CString
Following the same approach as GH-434, handle errors when parsing method
names into `CString`s at construction time. This allows invalid method
names to:

1. Not panic.
2. Properly surface as errors during boot.

the `extn` init infrastructure already passes through errors, so pass
back an error in this case.

This commit alters `method::Spec::new` to return
`Result<Self, ArtichokeError>` if `name` cannot be parsed into a
`CString`, which means it contains a NUL byte. Since we're there, this
commit also changes the `method::Spec` type to store the name as a
`Cow<'static, str>` like `class::Spec` and `module::Spec` already do.

With this commit, the only `unwrap`/`expect` remaining in
`artichoke-backend` is in the `Convert` implementation for `Array`.